pyproject.toml
[project]
authors = [
  {name = "Lengyue", email = "lengyue@lengyue.me"},
]
dependencies = [
  "click>=8.0.0",
  "pymongo[srv]>=4.6.1",
  "gradio>=4.7.1",
  "python-dotenv>=1.0.0",
  "loguru>=0.7.0",
  "librosa>=0.10.1",
  "boto3>=1.33.6",
  "pyannote.audio>=3.1.1"
]
description = "Annotate Audio Data"
license = {text = "Apache"}
name = "fish-data-engine"
readme = "README.md"
requires-python = ">=3.9"
version = "0.1.0"

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["fish_data_engine"]

[tool.ruff]
line-length = 120
